+++ modules/next/src/Plugin/PreviewUrlGeneratorManagerInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Drupal\next\Plugin;
+
+use Drupal\Component\Plugin\PluginManagerInterface;
+
+/**
+ * Defines an interface for the preview_url_generator plugin manager.
+ */
+interface PreviewUrlGeneratorManagerInterface extends PluginManagerInterface {
+
+}