+++ packages/next-drupal-query/tests/queries/formatter-only.ts
@@ -1,0 +1,16 @@
+import { QueryFormatter } from "../../src"
+
+type Input = {
+  firstName: string
+  lastName: string
+}
+
+type Output = {
+  name: string
+}
+
+export const formatter: QueryFormatter<Input, Output> = (input) => {
+  return {
+    name: `${input.firstName} ${input.lastName}`,
+  }
+}